+++ crates/ui/src/components/button/icon_button.rs
@@ -1,7 +1,8 @@
 use gpui::{AnyView, DefiniteLength};
 
 use crate::{prelude::*, ElevationIndex, SelectableButton};
-use crate::{ButtonCommon, ButtonLike, ButtonSize, ButtonStyle, IconName, IconSize};
+use super::button_like::{ButtonCommon, ButtonLike, ButtonSize, ButtonStyle};
+use crate::{IconName, IconSize};
 
 use super::button_icon::ButtonIcon;
