+++ crates/ui/src/components/label/highlighted_label.rs
@@ -62,7 +62,7 @@ impl LabelCommon for HighlightedLabel {
 
 pub fn highlight_ranges(
     text: &str,
-    indices: &Vec<usize>,
+    indices: &[usize],
     style: HighlightStyle,
 ) -> Vec<(Range<usize>, HighlightStyle)> {
     let mut highlight_indices = indices.iter().copied().peekable();
